requirements.txt
fire
pyserial
osprofile
requests